refactor: pages 폴더 정리
